feat: merging scripts to one
